[PyTorch][Static Runtime] Fast aliasing in select_tensor by manual borrowing
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68122

See code comments for details; in brief, we repurpose support
for borrowing `Tensor`s in `MaybeOwned` to make the `select_tensor`
output a borrowed IValue that we have to clean up manually.

If we have any other ops that always create a new reference to an
existing Tensor, we can easily apply this same optimization.
ghstack-source-id: 145687483

Differential Revision: [D32318674](https://our.internmc.facebook.com/intern/diff/D32318674/)