Merge index_{add,fill,copy,select} sampling
This was in the TODO list as the three operations are very similar.
Did this as one of them was failing in the noncontig tests and I wanted
to make sure that all of them were tested properly, as they all appear
in the derivative formulas of each other.

After this PR, these operations do pass the noncontiguous tests.

ghstack-source-id: d71b5b2e1bc1ca8c6e759c9ec6b8157bc14accda
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68184